move util fntn to libraries folder and update unit tests
