+++ apps/test/unit/utilsTest.js
@@ -21,7 +21,6 @@ const {
   normalize,
   stringifyQueryParams,
   getTabId,
-  extractTextFromCode,
 } = utils;
 
 describe('utils modules', () => {
@@ -751,164 +750,6 @@ describe('utils modules', () => {
 
       result = stripEncapsulatingDoubleQuotes('blah"');
       expect(result).to.equal('blah"');
-    });
-  });
-
-  describe('extractTextFromCode', () => {
-    it('extracts string literals with double quotes', () => {
-      const code = 'console.log("Hello World");';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('Hello World');
-    });
-
-    it('extracts string literals with single quotes', () => {
-      const code = "console.log('Hello World');";
-      const result = extractTextFromCode(code);
-      expect(result).to.include('Hello World');
-    });
-
-    it('extracts string literals with template literals', () => {
-      const code = 'console.log(`Hello World`);';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('Hello World');
-    });
-
-    it('extracts single-line comments', () => {
-      const code = '// This is a comment\nlet x = 1;';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('This is a comment');
-    });
-
-    it('extracts multi-line comments', () => {
-      const code = '/* This is a\nmulti-line comment */\nlet x = 1;';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('This is a');
-      expect(result).to.include('multi-line comment');
-    });
-
-    it('extracts identifiers that are 2 or more characters', () => {
-      const code = 'let myVar = 1;';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('let');
-      expect(result).to.include('myVar');
-    });
-
-    it('does not extract single-character identifiers', () => {
-      const code = 'let x = 1;';
-      const result = extractTextFromCode(code);
-      expect(result).to.not.include('x');
-      expect(result).to.include('let');
-      expect(result).to.not.include('1');
-      expect(result).to.not.include('==');
-    });
-
-    it('removes parentheses that affect tokenization', () => {
-      const code = 'if(artistList[i]) == artist';
-      const result = extractTextFromCode(code);
-      expect(result).to.not.include('(');
-      expect(result).to.not.include(')');
-      expect(result).to.include('artistList');
-      expect(result).to.not.include('if(artlistList');
-    });
-
-    it('removes brackets that affect tokenization', () => {
-      const code = 'arr[0] = "test";';
-      const result = extractTextFromCode(code);
-      expect(result).to.not.include('[');
-      expect(result).to.not.include(']');
-    });
-
-    it('removes braces that affect tokenization', () => {
-      const code = 'function test() { return "hello"; }';
-      const result = extractTextFromCode(code);
-      expect(result).to.not.include('{');
-      expect(result).to.not.include('}');
-    });
-
-    it('keeps asterisks for profanity obfuscation detection', () => {
-      const code = 'console.log("f*ck");';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('f*ck');
-    });
-
-    it('keeps @ symbol for profanity obfuscation detection', () => {
-      const code = 'console.log("sh@t");';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('sh@t');
-    });
-
-    it('keeps ! symbol for profanity obfuscation detection', () => {
-      const code = 'console.log("d!ck");';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('d!ck');
-    });
-
-    it('extracts text from complex code with multiple elements', () => {
-      const code = `
-        // Check user input
-        function validateInput(text) {
-          let result = "invalid";
-          return result;
-        }
-      `;
-      const result = extractTextFromCode(code);
-      expect(result).to.include('Check user input');
-      expect(result).to.include('invalid');
-      expect(result).to.include('validateInput');
-      expect(result).to.include('text');
-      expect(result).to.include('result');
-    });
-
-    it('handles empty string', () => {
-      const result = extractTextFromCode('');
-      expect(result).to.equal('');
-    });
-
-    it('handles null input', () => {
-      const result = extractTextFromCode(null);
-      expect(result).to.equal('');
-    });
-
-    it('handles undefined input', () => {
-      const result = extractTextFromCode(undefined);
-      expect(result).to.equal('');
-    });
-
-    it('extracts identifiers with numbers', () => {
-      const code = 'let num1 = 5; let var2 = 10;';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('num1');
-      expect(result).to.include('var2');
-    });
-
-    it('keeps underscores in identifiers for obfuscation detection', () => {
-      const code = 'let my_var = 5; let sh_t = 10;';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('my_var');
-      expect(result).to.include('sh_t');
-    });
-
-    it('removes semicolons and commas', () => {
-      const code = 'let a = 1, b = 2;';
-      const result = extractTextFromCode(code);
-      expect(result).to.not.include(';');
-      expect(result).to.not.include(',');
-    });
-
-    it('unescapes common escape sequences in strings', () => {
-      const code = 'let msg = "Hello\\nWorld";';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('Hello World');
-      expect(result).to.include('World');
-      expect(result).to.not.include('\\n');
-    });
-
-    it('extracts only "var" from "var x = 1;"', () => {
-      const code = 'var x = 1;';
-      const result = extractTextFromCode(code);
-      expect(result).to.include('var');
-      // x is only 1 character, should not be included
-      expect(result.split(/\s+/)).to.not.include('x');
     });
   });
 });
